update trip.export and trip.list views

+++ api/services/trip/src/providers/TripRepositoryProvider.ts
@@ -157,7 +157,7 @@ export class TripRepositoryProvider implements TripRepositoryInterface {
       text: `
       SELECT
         datetime::date as day,
-        sum(distance*seats)::int as distance,
+        sum(distance*seats)::bigint as distance,
         sum(seats+1)::int as carpoolers,
         count(*)::int as trip,
         '0'::int as trip_subsidized,
